+++ datasource-manage/src/main/java/com/github/cookbit/controller/DatasourceConfigController.java
@@ -14,20 +14,21 @@
  * limitations under the License.
  */
 
-package com.github.cookbit.db.model;
+package com.github.cookbit.controller;
 
-import lombok.Data;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.web.bind.annotation.RestController;
 
 /**
  * 数据源配置
  *
  * @auther 961374431@qq.com
- * @date 2022年03月03日
+ * @date 2022年03月06日
  */
-@Data
-public class DatasourceConfig {
-    private String driverClassName; // 驱动Class
-    private String url; // URL
-    private String username; // 用户名称
-    private String password; // 用户密码
+@RestController
+public class DatasourceConfigController {
+    private static final Logger logger = LoggerFactory.getLogger(DatasourceConfigController.class);
+
+
 }
